Remove the getSize() method of MailboxMembership (IMAP-184)
git-svn-id: https://svn.apache.org/repos/asf/james/imap/trunk@981472 13f79535-47bb-0310-9956-ffa450edef68

+++ message/src/main/java/org/apache/james/imap/encode/FetchResponseEncoder.java
@@ -277,7 +277,7 @@ public class FetchResponseEncoder extends AbstractChainedImapEncoder {
 
     private void encodeSize(ImapResponseComposer composer,
             final FetchResponse fetchResponse) throws IOException {
-        final Integer size = fetchResponse.getSize();
+        final Long size = fetchResponse.getSize();
         if (size != null) {
             // TODO: add method to composer
             composer.message("RFC822.SIZE");
